Fold the school search into the list it searches
The last inconsistency in how these controllers talk. schools-search reached
schools-list through three custom events, the only ones in the codebase, and
they existed solely to join these two - so a reader had to trace a protocol
across two files and two components to follow one search.

They were always one concern. The list already held searching and
searchResults, which existed only to receive those events, and its render
already reconciled a search against the collapse. Searching now sets that
state directly and calls render, and clearing is the clear() the list already
had.

Safe because the search panel is only ever rendered inside a schools-list, on
these two pages and nowhere else. Its markup stays a component; its targets
and actions now name the controller that was always driving them.

select-all-checkboxes is left alone - the support feedbacks page uses it too.

Four controllers on this page become three, and nothing in app/javascript
dispatches a custom event any more.
