+++ app/views/publish/course_wizards/steps/_schools.html.erb
@@ -34,9 +34,7 @@
       those the collapse threshold lets on screen. %>
   <div data-controller="schools-list"
        data-schools-list-collapse-after-value="<%= current_step.schools_collapse_threshold %>"
-       data-action="schools-search:filter->schools-list#filter
-                    schools-search:clear->schools-list#clear
-                    schools-search:restore->schools-list#restore">
+       data-action="keydown->schools-list#keydown">
     <%= render ::Publish::Schools::SearchPanelComponent.new(schools: current_step.schools, value: :uuid) %>
 
     <%= form.govuk_check_boxes_fieldset :school_uuids,
